requirements.txt
astroid==2.1.0
isort==4.3.4
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pyyaml==4.2b1
pylint==2.2.2
six==1.12.0
wrapt==1.11.1
